+++ app/src/main/java/org/simple/clinic/deeplink/DeepLinkEffect.kt
@@ -12,6 +12,8 @@ data class FetchPatient(val patientUuid: UUID) : DeepLinkEffect()
 
 data class NavigateToPatientSummary(val patientUuid: UUID) : DeepLinkEffect()
 
+data class NavigateToPatientSummaryWithTeleconsultLog(val patientUuid: UUID, val teleconsultRecordId: UUID?) : DeepLinkEffect()
+
 object ShowPatientDoesNotExist : DeepLinkEffect()
 
 // We will show this error when patient uuid is null