+++ app/src/main/java/org/simple/clinic/deeplink/DeepLinkModel.kt
@@ -6,10 +6,20 @@ import java.util.UUID
 
 @Parcelize
 data class DeepLinkModel(
-    val patientUuid: UUID?
+    val patientUuid: UUID?,
+    val teleconsultRecordId: UUID?,
+    val isLogTeleconsultDeepLink: Boolean
 ) : Parcelable {
 
   companion object {
-    fun default(patientUuid: UUID?) = DeepLinkModel(patientUuid = patientUuid)
+    fun default(
+        patientUuid: UUID?,
+        teleconsultRecordId: UUID?,
+        isLogTeleconsultDeepLink: Boolean
+    ) = DeepLinkModel(
+        patientUuid = patientUuid,
+        teleconsultRecordId = teleconsultRecordId,
+        isLogTeleconsultDeepLink = isLogTeleconsultDeepLink
+    )
   }
 }